13578 Move updateRecord override to package adapter
